2020.07.27 12:30  - Enum 코드 정리  - Enum 코드 메시지 > 프로퍼티로 이동(테스트 완료)  - 불필요한 import 제거

+++ src/main/java/com/nike/dnp/service/auth/AuthService.java
@@ -382,6 +382,14 @@ public class AuthService {
     public Auth delete(final Long authSeq) {
         log.info("AuthService.delete");
         final Auth auth = this.getById(authSeq);
+
+        if (auth.getSubAuths().size() > 0) {
+            throw new CodeMessageHandleException(
+                    ErrorEnumCode.AuthError.FAIL_DELETE.toString()
+                    , ErrorEnumCode.AuthError.FAIL_DELETE.getMessage()
+            );
+        }
+
         auth.delete();
         this.initAuthCache();
         this.remove(authSeq);
